:sparkles: Remove purchase token when subscription expires
The token is not going to be reused in this case, since the user will have to purchase a new subscription. The lack of a token can be used by the front-end to be able to tell that there is no subscription, instead of a paused or on hold subscription.